Update var type of Numer_Telefonu for Czytelnik entity in DB.

+++ Models/Rezerwacja.py
@@ -3,9 +3,6 @@ from dataclasses import dataclass
 
 @dataclass
 class Rezerwacja:
-    # Id_rezerwacji: int
     Id_rezerwacja: int
-    # Ksiazka_id_ksiazki : float
-    Id_ksiazka: float  # int??
-    # Czytelnik_id_czetelnika : int
+    Id_ksiazka: int
     Id_Czytelnik: int
